requirements.txt
atomicwrites==1.2.1
attrs==18.2.0
certifi==2018.11.29
chardet==3.0.4
Click==7.0
Flask==1.0.2
idna==2.7
itsdangerous==1.1.0
Jinja2==2.10
MarkupSafe==1.1.0
more-itertools==4.3.0
pluggy==0.8.0
psycopg2-binary==2.7.6.1
py==1.7.0
pymongo==3.7.2
pytest==4.0.1
requests==2.20.1
six==1.11.0
urllib3==1.24.1
Werkzeug==0.14.1
